move initToken&index to factory
